This patch changes the signing key used with SLK to a full key including public and private bits. This allows any developer to create a signed SLK build and it simplifies the build/debug process. You no longer need to run SkipVerification.bat after this patch. The patch changes all the public key values in the source code to match the new signing key.
This change also removes the auto-GACing of code during the build process. This was done to prevent build errors when building on Vista from a UAC permission restricted Command Prompt.

This change also changes the path used to find the Microsoft.Sharepoint.dll file to point into a directory in the enlistment. It also includes GACUtil in the Solution directory since GACUtil is not installed on Windows Server 2003 that does not have the .NET SDK or Visual Studio installed.

This change fixes a build break on my build box in SPPackageStore.cs. I suspect this build break is found on my machine due to the installation of VS 2008 Beta 2 and the corresponding patch to the .NET Runtime/Framework.

Upgrading to this release requires special steps.  These steps have been documented in the SLK\Solution\ReadMe.txt file.