This checkin converts SLK to Visual Studio 2008 format.  After this change, you must use Visual Studio 2008, not 2005, to compile SLK.
This change also fixed a minor build errors in SLK\Samples\ReportPages and SLK\Samples\WebService.

The readme file still refers to SharePoint 3.0 needing to be installed on the development computer.  NMake will successfully compile SLK on a computer that does not have SharePoint installed (such as a Vista or Windows XP box which cannot install SharePoint...), but I still found a few references to the SharePoint paths inside some non-critical build files so I did not remove this note at this time.