Work Item 14172: Convert SLK to MSIL so the same files work with x86 and x64.
This checkin converts the MRCI project (used to read Class Server LRM files) to /clrpure so that it generates MSIL.  MSIL is the .NET Intermediate Language which is "Just In Time" compiled for the CPU it's running on so it works on either x86 or x64 CPUs.  MRCI is the last code included in the SLK installation that is "CPU-Architecture-Dependent" so now we can have one set of installation files for both types of CPUs.