Update the SLKAssignmentBroker for the Tracking Services to synchronize with the database schema changes that came in with work items 14467 (Substitute a GUID for the long int in LearnerAssignmentItem.Id) and 14657 (Modify the SLK Schema so that PackageFormat.Id is Identity).
Modify the changes made as part of 14467 (GUID for LearnerAssignmentId) that changed the HTML Query Parameter initially defined in 14439 (Add support for the LearnerAssignmentId to HTML Content) from ?LearnerAssignmentId to ?LearnerAssignmentGuidId.  This change broke the Services and the Multiple Choice Question element.

Fix bug 15618 found in the initial CMI data model checkin where the link between ActivityAttempt and Interactions was omitted from TrackingDataContract.cs.