Bugfix 18090: readme.txt has incorrect instructions on compiling code.
Removed fmake.bat and all references to it in the documentation, replaced with references to nmake.
Removed FilterWithIni because this was only used by fmake.
Removed SdkDoc makefile command and associated files since this required downloading and installing another utility and the SdkDoc's are not being updated so they don't need rebuilding.